New Instructions
Added the GOTO and GOTOIF "instructions" (they're just essentially templates for JMP and JMPIF), NOT (which inverts a boolean expression), COPY (copies a stack value to another location), and ARG_TO (which pulls an argument to specified stack location. These are not normally accessible otherwise, as there was no way to access a parent scope's block's prepared values).